Major refactoring & cleanup: functional state
Fairly extensive refactoring and cleanup, including:

====== coverage_model.persistence.PersistenceLayer

MasterManager and ParameterManager classes:
* Centralize IO with master.hdf5 and <parameter>.hdf5 files respectively
* Provides:
** is_dirty(): utilizes internal _hmap dict and _dirty set in conjunction with overridden __setattr__ to keep track of which attributes have changed since the last flush
** flush(): uses is_dirty to only flush the dirty portions to disk
** _load(): reads existing hdf5 file and repopulates the object model

Integration of *Manager classes into PersistenceLayer
* Full migration to using the *Manager classes for all hdf5 file IO

Some optimizations with respect to expand_domain and write_brick (prevention of repeat calculations)

Domain expansion reworked to NOT be reliant on shared objects

====== coverage_model.parameter_values.AbstractParameterValue

Domain expansion reworked to NOT be reliant on shared objects

====== coverage_model.persistence.PersistentStorage

Changed logging within _calc_slices to 'trace' level

====== coverage_model.coverage.DomainSet

Simplified and removed shared objects

====== coverage_model.coverage.SimplexCoverage

First pass at "load" functionality.  It is functional, but likely to change slightly in terms of it's usage.

Some refactoring to utilize the changes to PersistenceLayer and DomainSet

====== coverage_model.test.examples

Added oneparamcov(): creates a coverage with ONLY a time parameter.  Intended to simplify testing of PersistenceLayer