Enhancements to BrickWriter* and Pidantic integration
Moved BrickWriterWorker to it's own file and provided a __main__ to allow it to operate as a standalone process.  ZeroMQ connection information is NOT configurable at this time.

Integrated Pidantic for management of pool of brick_writer processes when more than 1 brick_writer is requested

Added pidantic to setup.py as a dependency

Addressed some bugs within the BrickWriterDispatcher with respect to Failure handling and re-organization of work on stash & failure

Added pyzmq to buildout.cfg as a dependency